updated grpc node x,y types, added new grpc session position type, grpc updated link options to use int, fixed corehandlers handling of dup, fixed corexml type handling for link options, updated mobility config types to correlate with link options

+++ daemon/core/nodes/network.py
@@ -459,10 +459,10 @@ class CoreNetwork(CoreNetworkBase):
         netem = ["netem"]
         changed = max(changed, netif.setparam("delay", delay))
         if loss is not None:
-            loss = float(loss)
+            loss = int(loss)
         changed = max(changed, netif.setparam("loss", loss))
         if duplicate is not None:
-            duplicate = float(duplicate)
+            duplicate = int(duplicate)
         changed = max(changed, netif.setparam("duplicate", duplicate))
         changed = max(changed, netif.setparam("jitter", jitter))
         if not changed:
